+++ source/dotnet/Measurements.Client.UnitTests/MeasurementsClientTests.cs
@@ -63,14 +63,14 @@ public class MeasurementsClientTests
         Mock<IMeasurementsForDayResponseParser> measurementsForDayResponseParser)
     {
         // Arrange
-        var query = new GetAggregatedByMonthQuery("1234567890", new YearMonth(2025, 3));
+        var query = new GetMonthlyAggregateByDateQuery("1234567890", new YearMonth(2025, 3));
         var response = CreateResponse(HttpStatusCode.OK, TestAssets.MeasurementsAggregatedByDate);
         var httpClient = CreateHttpClient(response);
         var httpClientFactoryMock = CreateHttpClientFactoryMock(httpClient);
         var sut = new MeasurementsClient(httpClientFactoryMock.Object, measurementsForDayResponseParser.Object);
 
         // Act
-        var actual = (await sut.GetAggregatedByMonth(query, CancellationToken.None)).ToList();
+        var actual = (await sut.GetMonthlyAggregateByDateAsync(query, CancellationToken.None)).ToList();
 
         // Assert
         Assert.NotNull(actual);
@@ -85,18 +85,18 @@ public class MeasurementsClientTests
 
     [Theory]
     [AutoMoqData]
-    public async Task GetAggregatedMeasurementsByMonthAsync_WhenCalledDataIsMissing_ReturnsCompleteListOfMeasurementAggregations(
+    public async Task GetAggregatedMeasurementsByDateAsync_WhenCalledDataIsMissing_ReturnsCompleteListOfMeasurementAggregations(
         Mock<IMeasurementsForDayResponseParser> measurementsForDayResponseParser)
     {
         // Arrange
-        var query = new GetAggregatedByMonthQuery("1234567890", new YearMonth(2024, 10));
+        var query = new GetMonthlyAggregateByDateQuery("1234567890", new YearMonth(2024, 10));
         var response = CreateResponse(HttpStatusCode.OK, TestAssets.MeasurementsAggregatedByDateMissingMeasurements);
         var httpClient = CreateHttpClient(response);
         var httpClientFactoryMock = CreateHttpClientFactoryMock(httpClient);
         var sut = new MeasurementsClient(httpClientFactoryMock.Object, measurementsForDayResponseParser.Object);
 
         // Act
-        var actual = (await sut.GetAggregatedByMonth(query, CancellationToken.None)).ToList();
+        var actual = (await sut.GetMonthlyAggregateByDateAsync(query, CancellationToken.None)).ToList();
 
         // Assert
         Assert.NotNull(actual);
@@ -108,27 +108,50 @@ public class MeasurementsClientTests
 
     [Theory]
     [AutoMoqData]
-    public async Task GetAggregatedByYear_WhenCalledForMeasuredMeteringPoint_ReturnsListOfMeasurementAggregations(
+    public async Task GetAggregatedByMonth_WhenCalledForMeasuredMeteringPoint_ReturnsListOfMeasurementAggregations(
         Mock<IMeasurementsForDayResponseParser> measurementsForDayResponseParser)
     {
         // Arrange
-        var query = new GetAggregatedByYearQuery("1234567890", 2025);
+        var query = new GetYearlyAggregateByMonthQuery("1234567890", 2025);
         var response = CreateResponse(HttpStatusCode.OK, TestAssets.MeasurementsAggregatedByMonth);
         var httpClient = CreateHttpClient(response);
         var httpClientFactoryMock = CreateHttpClientFactoryMock(httpClient);
         var sut = new MeasurementsClient(httpClientFactoryMock.Object, measurementsForDayResponseParser.Object);
 
         // Act
-        var actual = (await sut.GetAggregatedByYear(query, CancellationToken.None)).ToList();
+        var actual = (await sut.GetYearlyAggregateByMonthAsync(query, CancellationToken.None)).ToList();
 
         // Assert
         Assert.NotNull(actual);
         Assert.Equal(12, actual.Count);
         Assert.True(actual.First().YearMonth == new YearMonth(2025, 1));
         Assert.True(actual.Last().YearMonth == new YearMonth(2025, 12));
-        Assert.False(actual.All(p => p.MissingValues));
         Assert.True(actual.All(p => p.Quality == Quality.Measured));
-        Assert.False(actual.All(p => p.ContainsUpdatedValues));
+        Assert.True(actual.All(p => p.Unit == Unit.kWh));
+    }
+
+    [Theory]
+    [AutoMoqData]
+    public async Task GetAggregatedByYear_WhenCalledForMeasuredMeteringPoint_ReturnsListOfMeasurementAggregations(
+        Mock<IMeasurementsForDayResponseParser> measurementsForDayResponseParser)
+    {
+        // Arrange
+        var query = new GetAggregateByYearQuery("1234567890");
+        var response = CreateResponse(HttpStatusCode.OK, TestAssets.MeasurementsAggregatedByYear);
+        var httpClient = CreateHttpClient(response);
+        var httpClientFactoryMock = CreateHttpClientFactoryMock(httpClient);
+        var sut = new MeasurementsClient(httpClientFactoryMock.Object, measurementsForDayResponseParser.Object);
+
+        // Act
+        var actual = (await sut.GetAggregateByYearAsync(query, CancellationToken.None)).ToList();
+
+        // Assert
+        Assert.NotNull(actual);
+        Assert.Equal(5, actual.Count);
+        Assert.Equal(2021, actual.First().Year);
+        Assert.Equal(2025, actual.Last().Year);
+        Assert.Equal(Quality.Measured, actual.First().Quality);
+        Assert.Equal(Quality.Estimated, actual.Last().Quality);
         Assert.True(actual.All(p => p.Unit == Unit.kWh));
     }