+++ source/dotnet/Measurements.WebApi.UnitTests/MeasurementsControllerTests.cs
@@ -1,5 +1,4 @@
 ﻿using System.Dynamic;
-using System.Net;
 using AutoFixture.Xunit2;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 using Energinet.DataHub.Measurements.Application.Exceptions;
@@ -53,7 +52,7 @@ public class MeasurementsControllerTests
         var jsonSerializer = new JsonSerializer();
         measurementsHandler
             .Setup(x => x.GetByPeriodAsync(It.IsAny<GetByPeriodRequest>()))
-            .ThrowsAsync(new MeasurementsNotFoundDuringPeriodException());
+            .ThrowsAsync(new MeasurementsNotFoundException());
         var sut = new MeasurementsController(measurementsHandler.Object, logger.Object, jsonSerializer);
 
         // Act
@@ -65,7 +64,7 @@ public class MeasurementsControllerTests
 
     [Theory]
     [AutoData]
-    public async Task GetByPeriodAsync_WhenMeasurementsUnknownError_ReturnsInternalServerError(
+    public async Task GetByPeriodAsync_WhenMeasurementsUnknownError_ThrowsException(
         GetByPeriodRequest request,
         Mock<IMeasurementsHandler> measurementsHandler,
         Mock<ILogger<MeasurementsController>> logger)
@@ -77,12 +76,8 @@ public class MeasurementsControllerTests
             .ThrowsAsync(new Exception());
         var sut = new MeasurementsController(measurementsHandler.Object, logger.Object, jsonSerializer);
 
-        // Act
-        var actual = await sut.GetByPeriodAsync(request);
-
-        // Assert
-        Assert.IsType<ObjectResult>(actual);
-        Assert.Equivalent(HttpStatusCode.InternalServerError, ((ObjectResult)actual).StatusCode);
+        // Act & Assert
+        await Assert.ThrowsAsync<Exception>(async () => await sut.GetByPeriodAsync(request));
     }
 
     [Theory]
@@ -94,7 +89,7 @@ public class MeasurementsControllerTests
     {
         // Arrange
         var jsonSerializer = new JsonSerializer();
-        var response = CreateMeasurementsAggregatedByDateResponse();
+        var response = CreateMeasurementsAggregatedResponse(MeasurementsAggregatedByDateResponse.Create);
         var expected = CreateExpectedMeasurementsAggregatedByDate();
         measurementsHandler
             .Setup(x => x.GetAggregatedByDateAsync(It.IsAny<GetAggregatedByDateRequest>()))
@@ -119,7 +114,7 @@ public class MeasurementsControllerTests
         var jsonSerializer = new JsonSerializer();
         measurementsHandler
             .Setup(x => x.GetAggregatedByDateAsync(It.IsAny<GetAggregatedByDateRequest>()))
-            .ThrowsAsync(new MeasurementsNotFoundDuringPeriodException());
+            .ThrowsAsync(new MeasurementsNotFoundException());
         var sut = new MeasurementsController(measurementsHandler.Object, logger.Object, jsonSerializer);
 
         // Act
@@ -131,7 +126,7 @@ public class MeasurementsControllerTests
 
     [Theory]
     [AutoData]
-    public async Task GetAggregatedByDateAsync_WhenMeasurementsUnknownError_ReturnsInternalServerError(
+    public async Task GetAggregatedByDateAsync_WhenMeasurementsUnknownError_ThenThrowsException(
         GetAggregatedByDateRequest request,
         Mock<IMeasurementsHandler> measurementsHandler,
         Mock<ILogger<MeasurementsController>> logger)
@@ -143,12 +138,8 @@ public class MeasurementsControllerTests
             .ThrowsAsync(new Exception());
         var sut = new MeasurementsController(measurementsHandler.Object, logger.Object, jsonSerializer);
 
-        // Act
-        var actual = await sut.GetAggregatedByDateAsync(request);
-
-        // Assert
-        Assert.IsType<ObjectResult>(actual);
-        Assert.Equivalent(HttpStatusCode.InternalServerError, ((ObjectResult)actual).StatusCode);
+        // Act & Assert
+        await Assert.ThrowsAsync<Exception>(async () => await sut.GetAggregatedByDateAsync(request));
     }
 
     [Theory]
@@ -160,7 +151,7 @@ public class MeasurementsControllerTests
     {
         // Arrange
         var jsonSerializer = new JsonSerializer();
-        var response = CreateMeasurementsAggregatedByMonthResponse();
+        var response = CreateMeasurementsAggregatedResponse(MeasurementsAggregatedByMonthResponse.Create);
         var expected = CreateExpectedMeasurementsAggregatedByMonth();
         measurementsHandler
             .Setup(x => x.GetAggregatedByMonthAsync(It.IsAny<GetAggregatedByMonthRequest>()))
@@ -185,7 +176,7 @@ public class MeasurementsControllerTests
         var jsonSerializer = new JsonSerializer();
         measurementsHandler
             .Setup(x => x.GetAggregatedByMonthAsync(It.IsAny<GetAggregatedByMonthRequest>()))
-            .ThrowsAsync(new MeasurementsNotFoundDuringPeriodException());
+            .ThrowsAsync(new MeasurementsNotFoundException());
         var sut = new MeasurementsController(measurementsHandler.Object, logger.Object, jsonSerializer);
 
         // Act
@@ -197,7 +188,7 @@ public class MeasurementsControllerTests
 
     [Theory]
     [AutoData]
-    public async Task GetAggregatedByMonthAsync_WhenMeasurementsUnknownError_ReturnsInternalServerError(
+    public async Task GetAggregatedByMonthAsync_WhenMeasurementsUnknownError_ThrowsException(
         GetAggregatedByMonthRequest request,
         Mock<IMeasurementsHandler> measurementsHandler,
         Mock<ILogger<MeasurementsController>> logger)
@@ -209,33 +200,83 @@ public class MeasurementsControllerTests
             .ThrowsAsync(new Exception());
         var sut = new MeasurementsController(measurementsHandler.Object, logger.Object, jsonSerializer);
 
+        // Act & Assert
+        await Assert.ThrowsAsync<Exception>(async () => await sut.GetAggregatedByMonthAsync(request));
+    }
+
+    [Theory]
+    [AutoData]
+    public async Task GetAggregatedByYearAsync_WhenMeasurementsExists_ReturnValidJson(
+        GetAggregatedByYearRequest request,
+        Mock<IMeasurementsHandler> measurementsHandler,
+        Mock<ILogger<MeasurementsController>> logger)
+    {
+        // Arrange
+        var jsonSerializer = new JsonSerializer();
+        var response = CreateMeasurementsAggregatedResponse(MeasurementsAggregatedByYearResponse.Create);
+        var expected = CreateExpectedMeasurementsAggregatedByYear();
+        measurementsHandler
+            .Setup(x => x.GetAggregatedByYearAsync(It.IsAny<GetAggregatedByYearRequest>()))
+            .ReturnsAsync(response);
+        var sut = new MeasurementsController(measurementsHandler.Object, logger.Object, jsonSerializer);
+
         // Act
-        var actual = await sut.GetAggregatedByMonthAsync(request);
+        var actual = (await sut.GetAggregatedByYearAsync(request) as OkObjectResult)!.Value!.ToString();
 
         // Assert
-        Assert.IsType<ObjectResult>(actual);
-        Assert.Equivalent(HttpStatusCode.InternalServerError, ((ObjectResult)actual).StatusCode);
+        Assert.Equal(expected, actual);
     }
 
-    private static GetMeasurementResponse CreateMeasurementResponse()
+    [Theory]
+    [AutoMoqData]
+    public async Task GetAggregatedByYearAsync_WhenMeasurementsDoNotExist_ReturnsNotFound(
+        GetAggregatedByYearRequest request,
+        Mock<IMeasurementsHandler> measurementsHandler,
+        Mock<ILogger<MeasurementsController>> logger)
+    {
+        // Arrange
+        var jsonSerializer = new JsonSerializer();
+        measurementsHandler
+            .Setup(x => x.GetAggregatedByYearAsync(It.IsAny<GetAggregatedByYearRequest>()))
+            .ThrowsAsync(new MeasurementsNotFoundException());
+        var sut = new MeasurementsController(measurementsHandler.Object, logger.Object, jsonSerializer);
+
+        // Act
+        var actual = await sut.GetAggregatedByYearAsync(request);
+
+        // Assert
+        Assert.IsType<NotFoundObjectResult>(actual);
+    }
+
+    [Theory]
+    [AutoData]
+    public async Task GetAggregatedByYearAsync_WhenMeasurementsUnknownError_ThrowsException(
+        GetAggregatedByYearRequest request,
+        Mock<IMeasurementsHandler> measurementsHandler,
+        Mock<ILogger<MeasurementsController>> logger)
+    {
+        // Arrange
+        var jsonSerializer = new JsonSerializer();
+        measurementsHandler
+            .Setup(x => x.GetAggregatedByYearAsync(It.IsAny<GetAggregatedByYearRequest>()))
+            .ThrowsAsync(new Exception());
+        var sut = new MeasurementsController(measurementsHandler.Object, logger.Object, jsonSerializer);
+
+        // Act & Assert
+        await Assert.ThrowsAsync<Exception>(async () => await sut.GetAggregatedByYearAsync(request));
+    }
+
+    private static MeasurementsResponse CreateMeasurementResponse()
     {
         var measurements = new List<MeasurementResult> { new(CreateMeasurementResult()) };
-        var response = GetMeasurementResponse.Create(measurements);
+        var response = MeasurementsResponse.Create(measurements);
         return response;
     }
 
-    private static GetMeasurementsAggregatedByDateResponse CreateMeasurementsAggregatedByDateResponse()
+    private static T CreateMeasurementsAggregatedResponse<T>(Func<List<AggregatedMeasurementsResult>, T> create)
     {
         var measurements = new List<AggregatedMeasurementsResult> { new(CreateAggregatedMeasurementResult()) };
-        var response = GetMeasurementsAggregatedByDateResponse.Create(measurements);
-        return response;
-    }
-
-    private static GetMeasurementsAggregatedByMonthResponse CreateMeasurementsAggregatedByMonthResponse()
-    {
-        var measurements = new List<AggregatedMeasurementsResult> { new(CreateAggregatedMeasurementResult()) };
-        var response = GetMeasurementsAggregatedByMonthResponse.Create(measurements);
-        return response;
+        return create(measurements);
     }
 
     private static string CreateExpectedMeasurementsByDate()
@@ -251,6 +292,11 @@ public class MeasurementsControllerTests
     private static string CreateExpectedMeasurementsAggregatedByMonth()
     {
         return """{"MeasurementAggregations":[{"YearMonth":"2023-09","Quantity":42,"Quality":"Measured","Unit":"kWh"}]}""";
+    }
+
+    private static string CreateExpectedMeasurementsAggregatedByYear()
+    {
+        return """{"MeasurementAggregations":[{"Year":2023,"Quantity":42,"Quality":"Measured","Unit":"kWh"}]}""";
     }
 
     private static ExpandoObject CreateMeasurementResult()