+++ source/geh_calculated_measurements/src/geh_calculated_measurements/common/infrastructure/__init__.py
@@ -1,20 +1,13 @@
-from geh_calculated_measurements.common.infrastructure.calculated_measurements.repository import (
-    Repository as CalculatedMeasurementsRepository,
+from geh_calculated_measurements.common.infrastructure.calculated_measurements_repository import (
+    CalculatedMeasurementsRepository,
 )
-from geh_calculated_measurements.common.infrastructure.current_measurements.repository import (
-    Repository as CurrentMeasurementsRepository,
-)
-from geh_calculated_measurements.common.infrastructure.database_definitions import (
-    CalculatedMeasurementsDatabaseDefinition,
-    CalculatedMeasurementsInternalDatabaseDefinition,
+from geh_calculated_measurements.common.infrastructure.current_measurements_repository import (
+    CurrentMeasurementsRepository,
 )
 from geh_calculated_measurements.common.infrastructure.spark_initializor import initialize_spark
 
 __all__ = [
-    "CalculatedMeasurementsInternalDatabaseDefinition",
-    "CalculatedMeasurementsDatabaseDefinition",
     "CalculatedMeasurementsRepository",
     "CurrentMeasurementsRepository",
     "initialize_spark",
-    "Table",
 ]