+++ source/geh_calculated_measurements/src/geh_calculated_measurements/missing_measurements_log/infrastructure/__init__.py
@@ -1,5 +1,11 @@
-from geh_calculated_measurements.missing_measurements_log.infrastructure.repository import Repository
+from geh_calculated_measurements.missing_measurements_log.infrastructure.electricity_market_repository import (
+    ElectricityMarketRepository,
+)
+from geh_calculated_measurements.missing_measurements_log.infrastructure.missing_measurements_log_repository import (
+    MissingMeasurementsLogRepository,
+)
 
 __all__ = [
-    "Repository",
+    "ElectricityMarketRepository",
+    "MissingMeasurementsLogRepository",
 ]